ai-proposal-agent/requirements.txt
streamlit>=1.32
langchain>=0.2.10
langchain-openai>=0.1.7
openai>=1.30.0
python-docx>=0.8.11
reportlab>=3.6.10
markdown>=3.4.4
pypandoc>=1.9.5
pytest>=7.0
pydantic>=1.10
python-multipart>=0.0.5
python-slugify>=8.0.4
pdfkit>=1.0.0
